Implemented Memory Card & A few GPU bug fixes

+++ Shaders/vertex.glsl
@@ -17,7 +17,7 @@ out vec4 UVs;
 flat out int attr;
 //flat out uint textureAttr[4];
 
-uniform ivec2 offset;
+uniform vec2 offset;
 uniform ivec2 drawingArea;
 
 const int TEXTURE_MODE_MASK = 0x1C;
@@ -40,10 +40,10 @@ void main() {
     * Converts VRAM coordinates (0; drawingArea.x, 0; drawingArea.y)
     * to OpenGL coordinates (-1;1, -1,1)
     */
-    float xPos = ((position.x / float(area.x)) * 2.0) - 1.0; 
+    float xPos = ((position.x / float(area.x)) * 2.0) - 1.0;
     
     // VRAM puts 0 at the top, OpenGL at the bottom..
-    float yPos = 1.0 - ((position.y / float(area.y)) * 2.0);
+    float yPos = 1.0 - (position.y / float(area.y)) * 2.0;
     
     // Set the final position
     gl_Position.xyzw = vec4(xPos, yPos, 0.0, 1.0);
